refactor(domain, exception): 전역 도메인 및 예외 처리 패키지 구조 재편성
프로젝트의 도메인 모델과 예외 처리 클래스들의 패키지 경로를 재편성함으로써, 코드의 일관성과 모듈화를 개선하고, 프로젝트의 구조를 더 명확하게 하였습니다. 주요 변경 내용은 다음과 같습니다:

1. `sidepair.global.domain`에서 사용되던 `BaseEntity`, `DomainException`, `ImageContentType` 클래스들을 `sidepair.domain` 패키지로 이동시켰습니다. 이는 도메인 관련 클래스들을 중앙화하여 관리의 효율성을 높이기 위함입니다.

2. 도메인 관련 예외 클래스들도 같은 방식으로 `sidepair.global.domain.exception`에서 `sidepair.domain.exception`으로 이동시켜, 예외 처리 구조 역시 도메인 중심으로 재구성하였습니다.

3. 이러한 변경을 통해, 도메인 모델과 관련된 예외 처리 로직이 같은 상위 패키지 내에 위치함으로써, 코드의 가독성 및 유지보수성이 향상되었습니다.

이번 구조 재편성 작업은 프로젝트의 확장성 및 유연성 측면에서 장기적인 이점을 제공할 것으로 기대됩니다.
