+++ src/main/java/sidepair/domain/BaseUpdatedTimeEntity.java
@@ -1,4 +1,4 @@
-package sidepair.global.domain;
+package sidepair.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.MappedSuperclass;
